backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.32.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.7
filelock==3.13.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
Levenshtein==0.26.1
MarkupSafe==2.1.5
mpmath==1.3.0
msrest==0.7.1
networkx==3.3
numpy==2.2.2
oauthlib==3.2.2
openai==1.60.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.45.1
